fix(enrolmentPage): refactor and move large elements to separate component
