+++ src/app/pages/auth/auth.module.ts
@@ -3,6 +3,7 @@ import { NgModule } from '@angular/core'
 import { IonicModule } from 'ionic-angular'
 
 import { PipesModule } from '../../shared/pipes/pipes.module'
+import { QRFormComponent } from './components/qr-form/qr-form.component'
 import { TokenFormComponent } from './components/token-form/token-form.component'
 import { EnrolmentPageComponent } from './containers/enrolment-page.component'
 import { AuthService } from './services/auth.service'
@@ -13,7 +14,7 @@ import { AuthService } from './services/auth.service'
     IonicModule.forRoot(EnrolmentPageComponent),
     PipesModule
   ],
-  declarations: [EnrolmentPageComponent, TokenFormComponent],
+  declarations: [EnrolmentPageComponent, TokenFormComponent, QRFormComponent],
   providers: [AuthService]
 })
 export class AuthModule {}
