+++ src/app/pages/auth/components/qr-form/qr-form.component.scss
@@ -1,0 +1,13 @@
+token-form {
+  .qr-code {
+    margin: auto;
+    padding-top: 3vh;
+    width: 170px;
+  }
+
+  .bottom {
+    position: fixed;
+    bottom: 0;
+    width: 100%;
+  }
+}
