+++ src/app/pages/auth/components/token-form/token-form.component.html
@@ -1,22 +1,32 @@
 <form id="tokenForm" [formGroup]="metaQRForm" (ngSubmit)="submitForm()">
-  <div class="registration">
-    <div class="text text-lg">{{ 'ENROL_REGISTRATION' | translate }}</div>
-    <div class="text text-justify">
-      {{ 'ENROL_REGISTRATION_TOKEN_DESC' | translate }}
-    </div>
-    <ion-input
-      required
-      formControlName="baseURL"
-      type="url"
-      placeholder="Base URL"
-      (ionFocus)="onFocus()"
-    ></ion-input>
-    <ion-input
-      required
-      formControlName="tokenName"
-      type="text"
-      placeholder="Token"
-      (ionFocus)="onFocus()"
-    ></ion-input>
+  <div class="text text-justify">
+    {{ 'ENROL_REGISTRATION_TOKEN_DESC' | translate }}
+  </div>
+  <ion-input
+    required
+    formControlName="baseURL"
+    type="url"
+    placeholder="Base URL"
+    (ionFocus)="onFocus()"
+  ></ion-input>
+  <ion-input
+    required
+    formControlName="tokenName"
+    type="text"
+    placeholder="Token"
+    (ionFocus)="onFocus()"
+    clearOnEdit="true"
+  ></ion-input>
+  <div class="bottom" padding *ngIf="!loading">
+    <button
+      form="tokenForm"
+      ion-button
+      class="bt bt--full"
+      round
+      full
+      type="submit"
+    >
+      {{ 'BTN_SUBMIT' | translate }}
+    </button>
   </div>
 </form>
