+++ src/app/pages/auth/components/token-form/token-form.component.ts
@@ -1,7 +1,10 @@
-import { Component, EventEmitter, Output } from '@angular/core'
+import { Component, EventEmitter, Input, Output } from '@angular/core'
 import { FormControl, FormGroup, Validators } from '@angular/forms'
 
-import { DefaultEnrolmentBaseURL } from '../../../../../assets/data/defaultConfig'
+import {
+  DefaultEnrolmentBaseURL,
+  DefaultMetaTokenURI
+} from '../../../../../assets/data/defaultConfig'
 import { isValidURL } from '../../../../shared/utilities/form-validators'
 
 @Component({
@@ -9,6 +12,9 @@ import { isValidURL } from '../../../../shared/utilities/form-validators'
   templateUrl: 'token-form.component.html'
 })
 export class TokenFormComponent {
+  @Input()
+  loading: boolean
+
   @Output()
   data: EventEmitter<any> = new EventEmitter<any>()
   @Output()
@@ -26,7 +32,11 @@ export class TokenFormComponent {
     const token = this.metaQRForm.get('tokenName').value.trim()
     if (!isValidURL(baseURL))
       return this.errors.emit({ error: { message: 'Enter a valid URL' } })
-    this.data.emit([baseURL, token])
+    this.data.emit(this.getURLFromToken(baseURL, token))
+  }
+
+  getURLFromToken(base, token) {
+    return base + DefaultMetaTokenURI + token
   }
 
   onFocus() {
